Create README.md
We are a group of tech geeks and designers. We have been benefiting from the Web 3.0 and blockchain innovation for a while now. But, Now is the time to contribute to its development through NFTs, Dapps, Games and other digital assets. We will be releasing collections of NFTs that will be utilized in our games. Join in and Enjoy.

The Web 3.0 is growing at a fast pace. We are going to use blockchain technology to build a robust gaming, Dapps, Non Fungible assets and Metaverse at Large. As a small team of designers, tech geeks and tech enthusiasts, we are beginning our journey with NFT.